partialize things more to add an avenue to delete remote and downloading movies

+++ app/models/video.rb
@@ -134,6 +134,7 @@ class Video < ActiveRecord::Base
   def redownload; end
 
   def delete_file!
+    return unless raw_file_path
     File.delete(raw_file_path)
   end
